Move custom type definition into shared/ directory
