+++ client/tsconfig.json
@@ -20,6 +20,6 @@
     "moduleResolution": "node",
     "esModuleInterop": true
   },
-  "include": ["src/**/*.ts", "../custom.d.ts"],
+  "include": ["src/**/*.ts", "../shared/custom.d.ts"],
   "exclude": ["node_modules"]
 }
